# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Course Administration API
  description: API para la administración de cursos universitarios
  version: 1.0.0
...
paths:
...
  /asistencia:
    get:
      tags:
        - Attendance
      summary: Obtener registros de asistencia
      description: Retorna registros de asistencia. Se puede filtrar por clase o por alumno.
      parameters:
        - in: query
          name: id_clase
          schema:
            type: integer
          description: ID de la clase para filtrar asistencia
        - in: query
          name: id_alumno
          schema:
            type: integer
          description: ID del alumno para filtrar asistencia
      responses:
        "200":
          description: Registros de asistencia obtenidos correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      attendance:
                        type: array
                        items:
                          $ref: "#/components/schemas/Attendance"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Attendance
      summary: Registrar asistencia
      description: >
        Registra la asistencia de un alumno a una clase.
        Valida el código QR (`code`) generado con SHA-256 a partir del `id_alumno`, `id_clase` y un secreto interno.
        También valida que el alumno esté activo y que su ubicación GPS esté dentro del radio permitido (300 m de FIUBA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceCreate"
      responses:
        "201":
          description: Asistencia registrada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Asistencia registrada correctamente
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /asistencia/{id}:
    patch:
      tags:
        - Attendance
      summary: Actualizar un registro de asistencia
      description: Modifica el campo `presente` de un registro de asistencia existente. Requiere autenticación.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del registro de asistencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - presente
              properties:
                presente:
                  type: boolean
                  description: Nuevo valor de presencia
                  example: true
      responses:
        "200":
          description: Asistencia actualizada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Asistencia actualizada con éxito
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Attendance
      summary: Eliminar un registro de asistencia
      description: Elimina un registro de asistencia por su ID. Requiere autenticación.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del registro de asistencia a eliminar
      responses:
        "200":
          description: Registro de asistencia eliminado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Registro de asistencia con ID 3 eliminado correctamente
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

# CLASES
...
  /marks:
    get:
      tags:
        - Marks
      summary: Obtener notas
      description: Retorna notas con filtros opcionales por alumno, evaluación, equipo, corrector o valor de nota.
      parameters:
        - in: query
          name: id_alumno
          schema:
            type: integer
          description: Filtrar por ID de alumno
        - in: query
          name: id_evaluacion
          schema:
            type: integer
          description: Filtrar por ID de evaluación
        - in: query
          name: id_equipo
          schema:
            type: integer
          description: Filtrar por ID de equipo
        - in: query
          name: corrector_nombre
          schema:
            type: string
          description: Filtrar por nombre del corrector
        - in: query
          name: nota
          schema:
            type: number
          description: Filtrar por valor exacto de nota
      responses:
        "200":
          description: Lista de notas obtenida correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      marks:
                        type: array
                        items:
                          $ref: "#/components/schemas/Mark"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Marks
      summary: Crear una nota
      description: Registra una nota para un alumno en una evaluación. Requiere autenticación.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkCreate"
      responses:
        "201":
          description: Nota creada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: nota creada correctamente
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /marks/{id_mark}:
    get:
      tags:
        - Marks
      summary: Obtener una nota por ID
      parameters:
        - in: path
          name: id_mark
          required: true
          schema:
            type: integer
          description: ID de la nota
      responses:
        "200":
          description: Nota encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      mark:
                        type: array
                        items:
                          $ref: "#/components/schemas/Mark"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - Marks
      summary: Actualizar parcialmente una nota
      description: Modifica uno o más campos de una nota existente. Requiere autenticación.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id_mark
          required: true
          schema:
            type: integer
          description: ID de la nota a actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPatch"
      responses:
        "200":
          description: Nota actualizada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Nota actualizada con éxito
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Marks
      summary: Eliminar una nota
      description: Elimina una nota por su ID. Requiere autenticación.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id_mark
          required: true
          schema:
            type: integer
          description: ID de la nota a eliminar
      responses:
        "200":
          description: Nota eliminada correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Nota con ID 7 eliminada correctamente
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

# MATERIALES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Attendance:
      type: object
      properties:
        id_asistencia:
          type: integer
          example: 42
        id_alumno:
          type: integer
          example: 7
        id_clase:
          type: integer
          example: 5
        presente:
          type: boolean
          example: true
    AttendanceCreate:
      type: object
      required:
        - id_alumno
        - id_clase
        - code
        - latitud
        - longitud
      properties:
        id_alumno:
          type: integer
          example: 7
        id_clase:
          type: integer
          example: 5
        code:
          type: string
          description: >
            Hash SHA-256 generado a partir de `{id_alumno}-{id_clase}-2026-secret`.
            Se obtiene escaneando el QR enviado por email al alumno.
          example: "a3f1c2d4e5..."
        latitud:
          type: number
          format: float
          description: Latitud GPS del alumno al momento de registrar asistencia
          example: -34.6177
        longitud:
          type: number
          format: float
          description: Longitud GPS del alumno al momento de registrar asistencia
          example: -58.3682
        presente:
          type: boolean
          default: true
          example: true
...
    Mark:
      type: object
      properties:
        id_nota:
          type: integer
          example: 15
        id_alumno:
          type: integer
          example: 7
        id_evaluacion:
          type: integer
          example: 1
        nota:
          type: number
          format: float
          example: 8.5
    MarkCreate:
      type: object
      required:
        - id_alumno
        - id_evaluacion
        - nota
      properties:
        id_alumno:
          type: integer
          example: 7
        id_evaluacion:
          type: integer
          example: 1
        nota:
          type: number
          format: float
          example: 8.5
        id_equipo:
          type: integer
          nullable: true
          example: 3
        corrector_nombre:
          type: string
          nullable: true
          example: Bruno
    MarkPatch:
      type: object
      description: Al menos uno de los campos debe estar presente.
      properties:
        id_alumno:
          type: integer
          example: 7
        id_evaluacion:
          type: integer
          example: 1
        id_equipo:
          type: integer
          example: 3
        nota:
          type: number
          format: float
          example: 9.0
        corrector_nombre:
          type: string
          example: Bruno
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        code:
          type: integer
          example: 404
        message:
          type: string
          example: Not Found
        description:
          type: string
          example: No existe un aviso con ID 99
  responses:
    BadRequest:
      description: Solicitud inválida o campos faltantes
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: 400
            message: Bad Request
            description: Los campos 'id_alumno' e 'id_clase' son obligatorios
    Unauthorized:
      description: Token de autenticación ausente o inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: 401
            message: Unauthorized
            description: Token inválido o expirado
    Forbidden:
      description: Acceso denegado (alumno inactivo o fuera del rango GPS)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: 403
            message: Forbidden
            description: Ubicación fuera del rango permitido
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: 404
            message: Not Found
            description: No existe un aviso con ID 99
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: 500
            message: Internal Server Error
            description: Unexpected error